Change radare2 version to latest This breaks iaito but oh well
- Version 5.5.4 works with iaito but segfaults in debugging when using .deb
package
- Version 5.6.8 works with debugging but but segfaults iaito (multithreaded? Broken api?)
- Version 5.5.4 compiled with sys/install.sh --install works with everything but
makes the docker layer huge unless I manually find all the bins and libs
and put them in the right place. If there are any changes this will
break which I don't want for an automated dockerfile.

Screw iaito for now we don't need it. Wait for them to catch up with the latest
radare build.