Ugh viml is the worst
Now we can open a jar file with vim, navigate to the class we want, and
automatically decompile it.

A little hacky, because we have to intercept the event before the buffer
is even written. Only two events are ever fired by Zip#ZipBrowseSelect,
both are pre-file loading.

We pull the file with the buffer name, extract it to a temp file,
decompile it with procyon, then set buffer contents to the decompiled
output. Then we jump back to the original buffer so vim can finish
opening the compressed contents into the new buffer.

This will open two windows. 1 with decompiled content. 2 compiled
garbage. Not ideal, but whatever.