ansible-lint gets confused if it's in a venv with symlinks
If you symlink ansible and ansible-lint outside of the venv, despite the
fact the venv links directly to the correct ansible inside the same
venv, it still thinks it's outside the venv.

We have to modify the $PATH to show ansible-lint that ansible is right
there.

Goofy, but they don't want you to mix a system ansible with a venv
ansible.